pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bookdatamaker"
version = "4.1.3"
description = "CLI tool for extracting text with DeepSeek OCR and generating datasets"
readme = "README.md"
requires-python = ">=3.10,<=3.13"
license = {text = "MIT"}
authors = [
    {name = "zwh20081", email = "zwh20081@solart.pro"}
]
dependencies = [
    "click",
    "httpx",
    "openai",
    "mcp",
    "pyarrow",
    "datasets",
    "python-dotenv",
    "rich",
    "aiofiles",
    "Pillow>=10.1.0",
    "tqdm",
    "PyMuPDF",
    "ebooklib",
    "beautifulsoup4",
    "python-pptx",
    "rapidfuzz",
]

[project.optional-dependencies]
local = [
    "transformers==4.46.3",
    "accelerate>=0.26.0",
    "tokenizers==0.20.3",
    "torch",
    "einops",
    "easydict",
    "addict",
    "img2pdf",
    "numpy",
]
local-flash = [
    "transformers==4.46.3",
    "tokenizers==0.20.3",
    "torch",
    "flash-attn",
    "einops",
    "easydict",
    "addict",
    "img2pdf",
    "numpy",
]
document = [
    "PyMuPDF",
    "ebooklib",
    "beautifulsoup4",
]
all = [
    "transformers==4.46.3",
    "tokenizers==0.20.3",
    "torch",
    "einops",
    "easydict",
    "addict",
    "img2pdf",
    "numpy",
    "PyMuPDF",
    "ebooklib",
    "beautifulsoup4",
]

dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "ruff",
    "mypy",
]

[project.scripts]
bookdatamaker = "bookdatamaker.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
